pyproject.toml
[project]
name = "variational-learning-benchmark"
version = "0.1.0"
description = "Variational Learning Benchmarks"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.12.0",
    "bayesian-torch>=0.5.0",
    "bert-score>=0.3.13",
    "datasets>=4.6.0",
    "einops>=0.8.2",
    "hydra-core>=1.3",
    "matplotlib>=3.10.8",
    "numpy>=2.4.2",
    "peft>=0.18.1",
    "pillow>=12.1.1",
    "rouge-score>=0.1.2",
    "sacrebleu>=2.6.0",
    "scikit-learn>=1.8.0",
    "scipy>=1.17.1",
    "tiktoken>=0.12.0",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "tqdm>=4.67.3",
    "transformers>=5.2.0",
    "wandb>=0.25.0",
]

[project.optional-dependencies]
ffcv = [
    "ffcv>=1.0.2",
    "cupy-cuda12x>=14.0.1",
    "numba>=0.64.0",
    "opencv-python-headless>=4.13.0.92",
]
mbr-metrics = [
    "bert-score>=0.3.13",
    "bleurt @ git+https://github.com/google-research/bleurt.git"
]
[tool.hatch.metadata]
allow-direct-references = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/vlbench", "src/vloptimizers"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
]
